+++ src/Query/Query.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 namespace Remorhaz\JSON\Path\Query;
 
 use Remorhaz\JSON\Path\Value\NodeValueListBuilder;
+use Throwable;
 use function call_user_func;
 use Remorhaz\JSON\Data\Value\NodeValueInterface;
-use Remorhaz\JSON\Path\Runtime\EvaluatorInterface;
 use Remorhaz\JSON\Path\Value\ValueListInterface;
 use Remorhaz\JSON\Path\Runtime\RuntimeInterface;
 
@@ -15,15 +15,12 @@ final class Query implements QueryInterface
 
     private $source;
 
-    private $callback;
+    private $callbackBuilder;
 
-    private $properties;
-
-    public function __construct(string $source, callable $callback, CapabilitiesInterface $properties)
+    public function __construct(string $source, CallbackBuilderInterface $callbackBuilder)
     {
         $this->source = $source;
-        $this->callback = $callback;
-        $this->properties = $properties;
+        $this->callbackBuilder = $callbackBuilder;
     }
 
     public function __invoke(NodeValueInterface $rootNode, RuntimeInterface $runtime): ValueListInterface
@@ -32,23 +29,42 @@ final class Query implements QueryInterface
             ->addValue($rootNode, 0)
             ->build();
 
-        return call_user_func(
-            $this->callback,
-            $input,
-            $runtime->getValueListFetcher(),
-            $runtime->getEvaluator(),
-            $runtime->getLiteralFactory(),
-            $runtime->getMatcherFactory(),
-        );
+        try {
+            return call_user_func(
+                $this
+                    ->callbackBuilder
+                    ->getCallback(),
+                $input,
+                $runtime->getValueListFetcher(),
+                $runtime->getEvaluator(),
+                $runtime->getLiteralFactory(),
+                $runtime->getMatcherFactory(),
+            );
+        } catch (Throwable $e) {
+            throw new Exception\QueryExecutionFailedException(
+                $this->source,
+                $this->callbackBuilder->getCallbackCode(),
+                $e,
+            );
+        }
     }
 
     public function getCapabilities(): CapabilitiesInterface
     {
-        return $this->properties;
+        return $this
+            ->callbackBuilder
+            ->getCapabilities();
     }
 
     public function getSource(): string
     {
         return $this->source;
+    }
+
+    public function getCallbackCode(): string
+    {
+        return $this
+            ->callbackBuilder
+            ->getCallbackCode();
     }
 }